+++ static/postcss.config.js
@@ -1,7 +1,20 @@
 module.exports = {
   plugins: {
-    autoprefixer: {
-      browsers: ["last 1 versions", "> 5%"]
-    }
-  }
+    autoprefixer: {},
+    "postcss-pxtorem": {
+      rootValue: 16,
+      propList: ["*"],
+    },
+    "postcss-extend-rule": {},
+    "postcss-preset-env": {
+      stage: 2,
+      browsers: [
+        "> 0.5%",
+        "last 2 versions",
+        "Firefox ESR",
+        "not dead",
+        "not op_mini all",
+      ],
+    },
+  },
 };